// src/features/people/types/PeopleItem.ts
export interface PeopleItem {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  films: string[];
  species: unknown[];
  vehicles: unknown[];
  starships: unknown[];
  created: string;
  edited: string;
  url: string;
}